Create BU_hiddenblock
This block is hidden from view to students and not easily editable by lecturers. It contains additional css for styling elements not on the main page (lessons and other activities/resources) and therefore cannot be styled by the main.css located in the menubar.
This block also contains the javascript for accordion element to minimise "uncaught..." or "accordion is not a..." error.