Update main.css
Added style for h3 font in Weekly title boxes. This is because I noticed a lecturer had changed her title to h3 instead of the default h1. However, h3 style is not accessible (low colour contrast) in the weekly header box.
The new style makes the font blue (if this is what lecturers are seeking) but a much darker blue to make it accessible.